# api/kube-jit/docs/swagger.yaml
basePath: /kube-jit-api
definitions:
...
  handlers.ClustersAndRolesResponse:
    properties:
      clusters:
        items:
          type: string
        type: array
      roles:
        items:
          $ref: '#/definitions/models.Roles'
        type: array
    type: object
...
  models.LoginResponse:
    properties:
      expiresIn:
        type: integer
      userData:
        $ref: '#/definitions/models.NormalizedUserData'
    type: object
...
  models.NormalizedUserData:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      provider:
        type: string
    type: object
...
  models.Roles:
    properties:
      name:
        type: string
    type: object
  models.SimpleMessageResponse:
    properties:
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  models.Team:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
info:
  contact: {}
  description: Self-service Kubernetes RBAC JIT Requests with Google/Azure/Github
    Oauth.
  title: Kube-JIT API
  version: "1.0"
paths:
  /k8s-callback:
    post:
      consumes:
      - application/json
      description: Used by the downstream Kubernetes controller to callback for status
        update. Validates the signed URL and updates the request status in the database.
        Returns a success message.
      parameters:
      - description: Callback payload (ticketID, status, message)
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Status updated successfully
          schema:
            $ref: '#/definitions/models.SimpleMessageResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.SimpleMessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.SimpleMessageResponse'
        "500":
          description: Failed to update request
          schema:
            $ref: '#/definitions/models.SimpleMessageResponse'
      summary: Kubernetes controller callback for status update
      tags:
      - k8s
...
  /kube-jit-api/approving-groups:
    get:
      consumes:
      - application/json
      description: |-
        Returns the list of platform approving groups for the authenticated user.
        Requires one or more cookies named kube_jit_session_<number> (e.g., kube_jit_session_0, kube_jit_session_1).
        Pass split cookies in the Cookie header, for example:
        -H "Cookie: kube_jit_session_0=${cookie_0};kube_jit_session_1=${cookie_1}"
        Note: Swagger UI cannot send custom Cookie headers due to browser security restrictions. Use curl for testing with split cookies.
      parameters:
      - description: 'Session cookies (multiple allowed, names: kube_jit_session_0,
          kube_jit_session_1, etc.)'
        in: header
        name: Cookie
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Team'
            type: array
        "401":
          description: 'Unauthorized: no token in session data'
          schema:
            $ref: '#/definitions/models.SimpleMessageResponse'
      summary: Get platform approving groups
      tags:
      - records
  /kube-jit-api/azure/profile:
    get:
      consumes:
      - application/json
      description: |-
        Returns the normalized Azure user profile for the authenticated user.
        Requires one or more cookies named kube_jit_session_<number> (e.g., kube_jit_session_0, kube_jit_session_1).
        Pass split cookies in the Cookie header, for example:
        -H "Cookie: kube_jit_session_0=${cookie_0};kube_jit_session_1=${cookie_1}"
        Note: Swagger UI cannot send custom Cookie headers due to browser security restrictions. Use curl for testing with split cookies.
      parameters:
      - description: 'Session cookies (multiple allowed, names: kube_jit_session_0,
          kube_jit_session_1, etc.)'
        in: header
        name: Cookie
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.NormalizedUserData'
        "401":
          description: 'Unauthorized: no token in session data'
          schema:
            $ref: '#/definitions/models.SimpleMessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.SimpleMessageResponse'
      summary: Get the logged in user's Azure profile
      tags:
      - azure
...
  /kube-jit-api/clusters-and-roles:
    get:
      consumes:
      - application/json
      description: |-
        Returns the list of clusters and roles available to the user.
        Requires one or more cookies named kube_jit_session_<number> (e.g., kube_jit_session_0, kube_jit_session_1).
        Pass split cookies in the Cookie header, for example:
        -H "Cookie: kube_jit_session_0=${cookie_0};kube_jit_session_1=${cookie_1}"
        Note: Swagger UI cannot send custom Cookie headers due to browser security restrictions. Use curl for testing with split cookies.
      parameters:
      - description: 'Session cookies (multiple allowed, names: kube_jit_session_0,
          kube_jit_session_1, etc.)'
        in: header
        name: Cookie
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: clusters and roles
          schema:
            $ref: '#/definitions/handlers.ClustersAndRolesResponse'
        "401":
          description: 'Unauthorized: no token in session data'
          schema:
            $ref: '#/definitions/models.SimpleMessageResponse'
      summary: Get available clusters and roles
      tags:
      - records
...
  /kube-jit-api/oauth/azure/callback:
    get:
      consumes:
      - application/json
      description: Handles the Azure OAuth callback, exchanges the code for an access
        token, fetches user info, sets session data, and returns normalized user data
        and expiration time.
      parameters:
      - description: Azure OAuth authorization code
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Normalized user data and expiration time
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Missing or invalid code
          schema:
            $ref: '#/definitions/models.SimpleMessageResponse'
        "403":
          description: Unauthorized domain
          schema:
            $ref: '#/definitions/models.SimpleMessageResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.SimpleMessageResponse'
      summary: Azure OAuth callback
      tags:
      - azure
...
swagger: "2.0"